@@ -1,11 +1,11 @@
 {
   "name": "mirador-imagecropper",
-  "version": "0.1.9",
-  "description": "Mirador 3 plugin to retrieve the image url for the selected area",
+  "version": "1.0.0-alpha.1",
+  "description": "Mirador 4 plugin to retrieve the image url for the selected area",
   "keywords": [
     "iiif",
     "mirador",
-    "mirador3",
+    "mirador4",
     "mirador-plugin",
     "react-component"
   ],
@@ -22,45 +22,68 @@
     "umd"
   ],
   "scripts": {
-    "build": "nwb build-react-component",
+    "build": "npm run build:umd",
+    "build:demo": "NODE_ENV=development webpack --mode=development",
+    "build:umd": "NODE_ENV=production webpack --mode=production",
+    "build:es": "mkdir -p es && cp -r src/* es && NODE_ENV=production MODULE_FORMAT=es npx babel es -d es",
+    "build:cjs": "mkdir -p lib && cp -r src/* lib && NODE_ENV=production MODULE_FORMAT=cjs npx babel lib -d lib",
     "clean": "nwb clean-module && nwb clean-demo",
     "format": "eslint --fix ./demo/src ./src",
     "lint": "eslint ./demo/src ./src",
     "prepublishOnly": "npm run build",
-    "start": "nwb serve-react-demo",
-    "test": "nwb test-react",
-    "test:coverage": "nwb test-react --coverage",
-    "test:watch": "nwb test-react --server"
+    "start": "NODE_ENV=development webpack serve --open"
   },
   "dependencies": {
-    "material-ui-image": "^3.3.2"
+    "@emotion/babel-plugin": "^11.11.0",
+    "@emotion/cache": "^11.11.0",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "mui-image": "^1.0.7"
   },
   "devDependencies": {
+    "@babel/cli": "^7.17.6",
+    "@babel/core": "^7.25.2",
+    "@babel/eslint-parser": "^7.25.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
+    "@babel/plugin-proposal-private-methods": "^7.18.6",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/plugin-transform-regenerator": "^7.16.7",
+    "@babel/plugin-transform-runtime": "^7.24.7",
+    "@babel/preset-env": "^7.25.3",
+    "@babel/preset-react": "^7.24.7",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.15",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^9.1.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsx-a11y": "^6.9.0",
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-simple-import-sort": "^12.1.1",
-    "mirador": "^3.3.0",
-    "nwb": "^0.25.2",
+    "html-loader": "^5.1.0",
+    "html-webpack-plugin": "^5.6.0",
+    "mirador": "^4.0.0-alpha.1",
     "prettier": "^3.3.3",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "webpack": "^5.93.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.7.4"
   },
   "peerDependencies": {
-    "@material-ui/core": "4.x",
-    "@material-ui/icons": "4.x",
-    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@mui/icons-material": "^5.x",
+    "@mui/material": "^5.x",
     "lodash": "4.x",
-    "mirador": "3.x",
+    "mirador": "^4.0.0-alpha.1",
     "openseadragon": "2.x",
-    "prop-types": "15.x",
-    "react": "16.x",
-    "react-dom": "16.x",
+    "react": "18.x",
+    "react-dom": "18.x",
     "react-rnd": "10.x",
     "redux-saga": "1.x",
     "reselect": "4.x"
