@@ -1,32 +1,35 @@
-import Box from "@material-ui/core/Box";
-import Button from "@material-ui/core/Button";
-import ButtonGroup from "@material-ui/core/ButtonGroup";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogTitle from "@material-ui/core/DialogTitle";
-import FormControl from "@material-ui/core/FormControl";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormLabel from "@material-ui/core/FormLabel";
-import Link from "@material-ui/core/Link";
-import Radio from "@material-ui/core/Radio";
-import RadioGroup from "@material-ui/core/RadioGroup";
-import Slider from "@material-ui/core/Slider";
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import Switch from "@material-ui/core/Switch";
-import TextField from "@material-ui/core/TextField";
-import Typography from "@material-ui/core/Typography";
-import Alert from "@material-ui/lab/Alert";
-import Image from "material-ui-image";
-import ns from "mirador/dist/es/src/config/css-ns";
-import ScrollIndicatedDialogContent from "mirador/dist/es/src/containers/ScrollIndicatedDialogContent";
-import PropTypes from "prop-types";
-import React, { useState } from "react";
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 
-import CopyToClipboard from "./dialog/CopyToClipboard";
-import RightsInformation from "./dialog/RightsInformation";
-import ShareButton from "./dialog/ShareButton";
-import { getAttributionString } from "./utils";
+import Alert from '@mui/material/Alert';
+import Box from '@mui/material/Box';
+import Button from '@mui/material/Button';
+import ButtonGroup from '@mui/material/ButtonGroup';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogTitle from '@mui/material/DialogTitle';
+import FormControl from '@mui/material/FormControl';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import FormGroup from '@mui/material/FormGroup';
+import FormLabel from '@mui/material/FormLabel';
+import Link from '@mui/material/Link';
+import Radio from '@mui/material/Radio';
+import RadioGroup from '@mui/material/RadioGroup';
+import Slider from '@mui/material/Slider';
+import Switch from '@mui/material/Switch';
+import TextField from '@mui/material/TextField';
+import Typography from '@mui/material/Typography';
+import { styled } from '@mui/material/styles';
+
+import ns from 'mirador/dist/es/src/config/css-ns';
+import ScrollIndicatedDialogContent from 'mirador/dist/es/src/containers/ScrollIndicatedDialogContent';
+
+import Image from 'mui-image';
+
+import CopyToClipboard from './dialog/CopyToClipboard';
+import RightsInformation from './dialog/RightsInformation';
+import ShareButton from './dialog/ShareButton';
+import { getAttributionString } from './utils';
 
 /** Converts the given absolute coordinates to relative ones with the given precision */
 const toRelativeCoordinates = ({ x, y, w, h }, width, height, precision) => ({
@@ -36,39 +39,37 @@ const toRelativeCoordinates = ({ x, y, w, h }, width, height, precision) => ({
   h: parseFloat(((h / height) * 100).toFixed(precision)),
 });
 
-const useStyles = makeStyles((theme) => ({
-  actions: {
-    justifyContent: "space-between",
-    flexWrap: "wrap",
-  },
-  actionButtons: {
-    flexWrap: "wrap",
-  },
-  alert: {
-    marginBottom: theme.spacing(1),
-  },
-  legend: {
-    width: "100%",
-    display: "flex",
-    justifyContent: "space-between",
-  },
-  optionsHeading: {
-    marginBottom: theme.spacing(1),
-    marginTop: theme.spacing(1),
-  },
-  previewHeading: {
-    marginBottom: theme.spacing(1),
-  },
-  previewImage: {
-    marginTop: theme.spacing(1),
-    marginBottom: theme.spacing(1),
-  },
-  previewLink: {
-    fontFamily: theme.typography.fontFamily ?? "sans-serif",
-  },
+const StyledAlert = styled(Alert)(({ theme }) => ({
+  marginBottom: theme.spacing(1),
+}));
+const StyledDialogActions = styled(DialogActions)({
+  justifyContent: 'space-between',
+  flexWrap: 'wrap',
+});
+const StyledButtonGroup = styled(ButtonGroup)({
+  flexWrap: 'wrap',
+});
+const StyledFormLabelLegend = styled(FormLabel)({
+  width: '100%',
+  display: 'flex',
+  justifyContent: 'space-between',
+});
+const StyledOptionsHeading = styled(Typography)(({ theme }) => ({
+  marginBottom: theme.spacing(1),
+  marginTop: theme.spacing(1),
+}));
+const StyledPreviewHeading = styled(Typography)(({ theme }) => ({
+  marginBottom: theme.spacing(1),
+}));
+const StyledPreviewImage = styled(Image)(({ theme }) => ({
+  marginTop: theme.spacing(1),
+  marginBottom: theme.spacing(1),
+}));
+const StyledPreviewLink = styled(Link)(({ theme }) => ({
+  fontFamily: theme.typography.fontFamily ?? 'sans-serif',
 }));
 
-const supportsClipboard = "clipboard" in navigator;
+const supportsClipboard = 'clipboard' in navigator;
 
 /** Renders the dialog where some IIIF parameters can be defined */
 const CroppingDialog = ({
@@ -84,43 +85,30 @@ const CroppingDialog = ({
   updateConfig,
   viewType,
 }) => {
-  const {
-    active,
-    dialogOpen,
-    enabled,
-    roundingPrecision,
-    showRightsInformation,
-  } = config;
+  const { active, dialogOpen, enabled, roundingPrecision, showRightsInformation } = config;
   const [copiedToClipboard, setCopiedToClipboard] = useState(false);
   const [mirrored, setMirrored] = useState(false);
-  const [quality, setQuality] = useState("default");
+  const [quality, setQuality] = useState('default');
   const [rotation, setRotation] = useState(0);
   const [size, setSize] = useState(100);
-  const {
-    actions,
-    actionButtons,
-    alert,
-    legend,
-    optionsHeading,
-    previewHeading,
-    previewImage,
-    previewLink,
-  } = useStyles();
   if (
     !enabled ||
     !active ||
     !dialogOpen ||
     !currentCanvas ||
-    viewType !== "single" ||
+    viewType !== 'single' ||
     !imageCoordinates
   ) {
     return null;
   }
+
+  /** Close imagecropper dialog */
   const closeDialog = () =>
     updateConfig({
       ...config,
       dialogOpen: false,
     });
+
   const attribution = getAttributionString(requiredStatement);
   const canvasWidth = currentCanvas.getWidth();
   const canvasHeight = currentCanvas.getHeight();
@@ -131,43 +119,39 @@ const CroppingDialog = ({
     roundingPrecision,
   );
   const region = `pct:${x},${y},${w},${h}`;
-  const mirror = mirrored ? "!" : "";
+  const mirror = mirrored ? '!' : '';
   const imageUrl = `${imageServiceIds[0]}/${region}/pct:${size}/${mirror}${rotation}/${quality}.jpg`;
+  /**
+   * Generates a URL for the preview image based on the given width.
+   * @param {number} width - The width of the preview image.
+   * @returns {string} The generated URL for the preview image.
+   */
   const getPreviewUrl = (width) =>
     `${imageServiceIds[0]}/${region}/${width},/${mirror}${rotation}/${quality}.jpg`;
-  /* The aspect ratio, which depends on the rotation, is only relevant for the preview image */
-  let aspectRatio = 1;
-  if (imageCoordinates.h > 0 && [0, 180].includes(rotation)) {
-    aspectRatio = imageCoordinates.w / imageCoordinates.h;
-  } else if (imageCoordinates.w > 0 && [90, 270].includes(rotation)) {
-    aspectRatio = imageCoordinates.h / imageCoordinates.w;
-  }
+
   return (
     <Dialog
-      container={document.querySelector(`#${containerId} .${ns("viewer")}`)}
+      container={document.querySelector(`#${containerId} .${ns('viewer')}`)}
       fullWidth
       maxWidth="sm"
       onClose={closeDialog}
       open={dialogOpen}
       scroll="paper"
     >
-      <DialogTitle disableTypography>
+      <DialogTitle>
         <Typography variant="h4">
-          <Box fontWeight="fontWeightBold">
-            {t("imageCropper.linkToSelectedRegion")}
-          </Box>
+          <Box fontWeight="fontWeightBold">{t('imageCropper.linkToSelectedRegion')}</Box>
         </Typography>
       </DialogTitle>
       <ScrollIndicatedDialogContent dividers>
         {copiedToClipboard && (
-          <Alert
-            className={alert}
-            closeText={t("imageCropper.close")}
+          <StyledAlert
+            closeText={t('imageCropper.close')}
             onClose={() => setCopiedToClipboard(false)}
             severity="success"
           >
-            {t("imageCropper.copiedToClipboard")}
-          </Alert>
+            {t('imageCropper.copiedToClipboard')}
+          </StyledAlert>
         )}
         <TextField
           fullWidth
@@ -189,38 +173,29 @@ const CroppingDialog = ({
           value={imageUrl}
           variant="outlined"
         />
-        <Typography className={optionsHeading} variant="h5">
-          {t("imageCropper.options")}
-        </Typography>
+        <StyledOptionsHeading variant="h5">{t('imageCropper.options')}</StyledOptionsHeading>
         <FormControl component="fieldset" fullWidth>
-          <FormLabel component="legend" className={legend}>
-            {t("imageCropper.size")} <span>{size}%</span>
-          </FormLabel>
+          <StyledFormLabelLegend component="legend">
+            {t('imageCropper.size')} <span>{size}%</span>
+          </StyledFormLabelLegend>
           <Slider min={1} onChange={(_evt, s) => setSize(s)} value={size} />
         </FormControl>
         <FormControl component="fieldset">
-          <FormLabel component="legend">{t("imageCropper.rotation")}</FormLabel>
+          <FormLabel component="legend">{t('imageCropper.rotation')}</FormLabel>
           <RadioGroup
-            aria-label={t("imageCropper.rotation")}
+            aria-label={t('imageCropper.rotation')}
             name="rotation"
             onChange={(evt) => setRotation(parseInt(evt.target.value, 10))}
             row
             value={rotation}
           >
             {[0, 90, 180, 270].map((r) => (
-              <FormControlLabel
-                control={<Radio />}
-                key={`${r}°`}
-                label={`${r}°`}
-                value={r}
-              />
+              <FormControlLabel control={<Radio />} key={`${r}°`} label={`${r}°`} value={r} />
             ))}
           </RadioGroup>
         </FormControl>
         <FormControl component="fieldset" fullWidth>
-          <FormLabel component="legend">
-            {t("imageCropper.reflection")}
-          </FormLabel>
+          <FormLabel component="legend">{t('imageCropper.reflection')}</FormLabel>
           <FormGroup row>
             <FormControlLabel
               control={
@@ -230,22 +205,20 @@ const CroppingDialog = ({
                   onChange={(evt) => setMirrored(evt.target.checked)}
                 />
               }
-              label={t("imageCropper.mirror")}
+              label={t('imageCropper.mirror')}
             />
           </FormGroup>
         </FormControl>
         <FormControl component="fieldset">
-          <FormLabel component="legend">
-            {t("imageCropper.quality.label")}
-          </FormLabel>
+          <FormLabel component="legend">{t('imageCropper.quality.label')}</FormLabel>
           <RadioGroup
-            aria-label={t("imageCropper.quality.label")}
+            aria-label={t('imageCropper.quality.label')}
             name="quality"
             onChange={(evt) => setQuality(evt.target.value)}
             row
             value={quality}
           >
-            {["default", "color", "gray", "bitonal"].map((q) => (
+            {['default', 'color', 'gray', 'bitonal'].map((q) => (
               <FormControlLabel
                 control={<Radio />}
                 key={q}
@@ -255,28 +228,16 @@ const CroppingDialog = ({
             ))}
           </RadioGroup>
         </FormControl>
-        <Typography className={previewHeading} variant="h5">
-          {t("imageCropper.preview.label")}
-        </Typography>
-        <Link
-          className={previewLink}
-          href={imageUrl}
-          rel="noopener"
-          target="_blank"
-        >
-          {t("imageCropper.preview.link")}
-        </Link>
-        <Image
-          aspectRatio={aspectRatio}
-          className={previewImage}
-          color="transparent"
-          src={getPreviewUrl(500)}
-        />
+        <StyledPreviewHeading variant="h5">{t('imageCropper.preview.label')}</StyledPreviewHeading>
+        <StyledPreviewLink href={imageUrl} rel="noopener" target="_blank">
+          {t('imageCropper.preview.link')}
+        </StyledPreviewLink>
+        <StyledPreviewImage bgColor="transparent" src={getPreviewUrl(500)} />
         {showRightsInformation && <RightsInformation t={t} rights={rights} />}
       </ScrollIndicatedDialogContent>
-      <DialogActions className={actions}>
-        <ButtonGroup className={actionButtons}>
-          {["envelope", "facebook", "pinterest", "x", "whatsapp"].map((p) => (
+      <StyledDialogActions>
+        <StyledButtonGroup>
+          {['envelope', 'facebook', 'pinterest', 'x', 'whatsapp'].map((p) => (
             <ShareButton
               attribution={attribution}
               imageUrl={imageUrl}
@@ -287,18 +248,18 @@ const CroppingDialog = ({
               title={t(`imageCropper.share.${p}`)}
             />
           ))}
-        </ButtonGroup>
+        </StyledButtonGroup>
         <Button color="primary" onClick={closeDialog}>
-          {t("imageCropper.close")}
+          {t('imageCropper.close')}
         </Button>
-      </DialogActions>
+      </StyledDialogActions>
     </Dialog>
   );
 };
 
 CroppingDialog.defaultProps = {
   currentCanvas: undefined,
-  label: "",
+  label: '',
   requiredStatement: [],
   rights: [],
 };
