@@ -1,24 +1,23 @@
-import Link from "@material-ui/core/Link";
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import Alert from "@material-ui/lab/Alert";
-import PropTypes from "prop-types";
-import React from "react";
+import React from 'react';
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    marginTop: theme.spacing(2),
-  },
+import Alert from '@mui/material/Alert';
+import Link from '@mui/material/Link';
+import { styled } from '@mui/material/styles';
+
+import PropTypes from 'prop-types';
+
+const StyledAlert = styled(Alert)(({ theme }) => ({
+  marginTop: theme.spacing(2),
 }));
 
 /** Renders the rights information defined in the used manifest */
 const RightsInformation = ({ rights, t }) => {
-  const { root } = useStyles();
   if (!rights.length) {
     return null;
   }
   return (
-    <Alert className={root} severity="warning">
-      <span>{t("imageCropper.noteRights", { count: rights.length })}: </span>
+    <StyledAlert severity="warning">
+      <span>{t('imageCropper.noteRights', { count: rights.length })}: </span>
       {rights.length === 1 ? (
         <Link href={rights[0]} rel="noopener" target="_blank">
           {rights[0]}
@@ -34,7 +33,7 @@ const RightsInformation = ({ rights, t }) => {
           ))}
         </ul>
       )}
-    </Alert>
+    </StyledAlert>
   );
 };
 
