@@ -1,13 +1,15 @@
-import FacebookIcon from "@material-ui/icons/Facebook";
-import MailIcon from "@material-ui/icons/Mail";
-import PinterestIcon from "@material-ui/icons/Pinterest";
-import WhatsAppIcon from "@material-ui/icons/WhatsApp";
-import { MiradorMenuButton } from "mirador/dist/es/src/components/MiradorMenuButton";
-import PropTypes from "prop-types";
-import React from "react";
+import React from 'react';
 
-import XIcon from "../icons/XIcon";
-import { getShareLink } from "../utils";
+import FacebookIcon from '@mui/icons-material/Facebook';
+import MailIcon from '@mui/icons-material/Mail';
+import PinterestIcon from '@mui/icons-material/Pinterest';
+import WhatsAppIcon from '@mui/icons-material/WhatsApp';
+
+import { MiradorMenuButton } from 'mirador/dist/es/src/components/MiradorMenuButton';
+import PropTypes from 'prop-types';
+import XIcon from '../icons/XIcon';
+
+import { getShareLink } from '../utils';
 
 const iconMapping = {
   envelope: MailIcon,
@@ -18,29 +20,11 @@ const iconMapping = {
 };
 
 /** Renders a button for sharing the given content on one of the supported providers */
-const ShareButton = ({
-  attribution,
-  imageUrl,
-  label,
-  provider,
-  thumbnailUrl,
-  title,
-}) => {
-  const link = getShareLink(
-    attribution,
-    imageUrl,
-    label,
-    provider,
-    thumbnailUrl,
-  );
+const ShareButton = ({ attribution, imageUrl, label, provider, thumbnailUrl, title }) => {
+  const link = getShareLink(attribution, imageUrl, label, provider, thumbnailUrl);
   const ProviderIcon = iconMapping[provider];
   return (
-    <MiradorMenuButton
-      aria-label={title}
-      href={encodeURI(link)}
-      rel="noopener"
-      target="_blank"
-    >
+    <MiradorMenuButton aria-label={title} href={encodeURI(link)} rel="noopener" target="_blank">
       <ProviderIcon />
     </MiradorMenuButton>
   );
