@@ -5,8 +5,8 @@ const getAttributionString = (requiredStatement) => {
   }
   const initial = requiredStatement.shift();
   return requiredStatement.reduce(
-    (acc, current) => `${acc}, ${current.values.join(", ")}`,
-    initial.values.join(", "),
+    (acc, current) => `${acc}, ${current.values.join(', ')}`,
+    initial.values.join(', '),
   );
 };
 
@@ -17,15 +17,15 @@ const getShareLink = (attribution, imageUrl, label, provider, thumbnailUrl) => {
     text += ` (${attribution})`;
   }
   switch (provider) {
-    case "envelope":
+    case 'envelope':
       return `mailto:?subject=${text}&body=${text}: ${imageUrl}`;
-    case "facebook":
+    case 'facebook':
       return `https://www.facebook.com/sharer/sharer.php?title=${text}&u=${imageUrl}`;
-    case "pinterest":
+    case 'pinterest':
       return `http://pinterest.com/pin/create/bookmarklet/?url=${imageUrl}&description=${text}&media=${thumbnailUrl}`;
-    case "whatsapp":
+    case 'whatsapp':
       return `whatsapp://send?text=${text}: ${imageUrl}`;
-    case "x":
+    case 'x':
       return `https://x.com/intent/post?text=${
         text.length > 60 ? `${text.substring(0, 60)}...` : text
       }&url=${imageUrl}&hashtags=iiif`;
